created a directory to hold HEAT UI components and moved Case Summary components to that directory
