+++ heat-stack/app/components/ui/heat/CaseSummaryComponents/CurrentHeatingSystem.tsx
