+++ heat-stack/app/components/ui/heat/CaseSummaryComponents/Graphs.tsx
